refactor: organizes Toast exports in their respective folder
