--- src/components/ui/toast/toast.d.ts
@@ -1,13 +1,0 @@
-declare module 'toast' {
-  export interface toast {
-    protocol?: string;
-    hostname?: string;
-    pathname?: string;
-  }
-}
-
-export interface toast {
-  protocol?: string;
-  hostname?: string;
-  pathname?: string;
-}
