+++ src/index.ts
@@ -98,10 +98,7 @@ export * from './components/ui/tab';
 export * from './components/ui/tag';
 export * from './components/ui/text-style';
 export * from './components/ui/timeline';
+export * from './components/ui/toast';
 
 // Types
 export * from './types';
-
-export { Toast as $toast } from './components/ui/toast';
-export { default as Toast } from './components/ui/toast/Toast.vue';
-export { default as Badge } from './components/ui/badge/Badge.vue';
